refactor(cli): extract light-mode detection into hermes_cli/light_mode
cli.py had grown past 14.9k lines. Move the self-contained light/dark
terminal detection + skin-color remap cluster (hex->ANSI conversion,
luminance, OSC 11 background probe, env/COLORFGBG detection, the
dark->light remap table, and the SkinConfig.get_color hook) into a new
pure module, hermes_cli/light_mode.py.

cli.py re-exports the helpers so every existing call site
(cli._hex_to_ansi, cli._detect_light_mode, _SkinAwareAnsi, the banner
color paths) keeps working unchanged, and retains the two import-time
side effects (_install_skin_light_mode_hook() + the interactive OSC 11
prime) so module-load behaviour is byte-identical.

The mutable _LIGHT_MODE_CACHE global now lives in the new module, so the
light-mode tests target hermes_cli.light_mode (importing cli first to
trigger the hook install the SkinConfigHook tests assert on).

Net: cli.py -246 lines; behaviour unchanged (20 light-mode + 13 skin
integration tests pass).
